+++ src/requiredAlgorithm.ts
@@ -1,4 +1,0 @@
-/*!
- * Copyright (c) 2023 Digital Bazaar, Inc. All rights reserved.
- */
-export const requiredAlgorithm = ['P-256', 'P-384']